pyproject.toml
[project]
name = "workllm"
version = "0.1.0"
description = "LLM-powered productivity toolkit for engineers"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "click>=8.1.8",
    "chromadb>=0.4.15",
    "ollama>=0.1.7",
    "openai>=1.12.0",
    "python-dotenv>=1.0.0",
    "python-magic>=0.4.27",
    "pytest>=8.3.4",
    "requests>=2.31.0",
    "rich>=13.7.0",
    "typing-extensions>=4.9.0",
    "unstructured>=0.12.0",
    "pdfminer>=20191125",
    "pdfminer-six>=20240706",
    "pi-heif>=0.21.0",
    "unstructured-inference>=0.8.6",
    "pdf2image>=1.17.0",
    "unstructured-pytesseract>=0.3.13",
    "pygithub>=2.3.0",
]

[project.scripts]
workllm = "workllm.cli:cli"

[build-system]
requires = ["setuptools>=65.5.1"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["workllm"]
include-package-data = true
package-dir = {"" = "workllm"}
